refactor: View 분리 및 네이밍 수정, 개행정리
