+++ Calculator/Calculator/Model/Operator.swift
@@ -1,5 +1,5 @@
 //  Calculator - Operator.swift
-//  created by vetto on 2023/01/27
+//  created by 리지, vetto, Andrew on 2023/01/27
 
 enum Operator: Character, CaseIterable, CalculateItem {
     case add = "+"
