+++ Calculator/Calculator/Protocol/CalculateItemProtocol.swift
@@ -1,5 +1,5 @@
 //  Calculator - CalculateItemProtocol.swift
-//  created by vetto on 2023/01/25
+//  created by 리지, vetto, Andrew on 2023/01/25
 
 protocol CalculateItem {}
