+++ Calculator/Calculator/Resources/Assets.xcassets/Contents.json
