+++ Calculator/Calculator/Resources/Info.plist
